+++ src/render/widget.rs
@@ -25,8 +25,6 @@ use sdl2::event::Event;
 use sdl2::pixels::Color;
 use std::collections::HashMap;
 
-pub trait ConfigConvenience {}
-
 /// This trait is shared by all `Widget` objects that have a presence on the screen.  Functions that
 /// must be implemented are documented in the trait.
 ///