+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Changed build process to load in Ubuntu libraries for SDL2 build
 - Updated Ubuntu documentation thanks to @alexislozano
 - Added handler for all other events in a catch-all function for `Widget`s (#227)
+- Fixed travis-ci build for OS X builds
 
 ## 0.4.12