+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added handler for all other events in a catch-all function for `Widget`s (#227)
 - Fixed travis-ci build for OS X builds
 - Modified `Widget`s so that they display the events sent to them that are unhandled.
+- Added `Layout` class, back ported from the original class, greatly simplified. (#230)
 
 ## 0.4.12